New snapshot iterator and position format
The iterator supports snapshots based on range and is limited to tables
which have an ordinal number (integer) as primary key. If these conditions are
not met, validation will fail.

Support for multiple tables is builtin. Each table gets a dedicated worker.
Positional data keeps track of the progress of all tables, not just the one which the record
is related to.

Resuming a snapshot is supported, if the snapshot positional data requires it.

This work is not integrated into the CDC or LongPoll iterator and the new position type
is not compatible. This will work will follow.