New snapshot iterator and position format (#132)
* New snapshot iterator and position format

The iterator supports snapshots based on range and is limited to tables
which have an ordinal number (integer) as primary key. If these conditions are
not met, validation will fail.

Support for multiple tables is builtin. Each table gets a dedicated worker.
Positional data keeps track of the progress of all tables, not just the one which the record
is related to.

Resuming a snapshot is supported, if the snapshot positional data requires it.

This work is not integrated into the CDC or LongPoll iterator and the new position type
is not compatible. This will work will follow.

* remove done channel

* position test; add header and fix linter

* fmt

* tests; adjust pg config for testing

* bump go

* add stringer

* pass db pool as param to iterator

* attach methods to receiver; rename fetcherworker to fetchworker; remove usage of sprint

* add validation for key not being numeric

* cr changes

* remove fmt, one type of params is enough

* lint

* simplifications

* include context in tomb

* return snapshot position explicitly in fetcher

* fix linter warning

* set correct position type

* make test assertion more stable

* document LSN

* rename snapshot to snapshots

---------

Co-authored-by: Lovro Mažgon <lovro.mazgon@gmail.com>