Add default logrepl autoCleanup mode to remove replication slots and publications
The new option `logrepl.autoCleanup` is not only valid in the context of logical replication.
The default is set to true and will remove the replication slot and publication when the connector is deleted.

This change is a departure from the current implementation, in the following areas:

1. Replication slots are not `temporary` any longer and will persist through pipeline state changes.
2. An implication of 1) the slot will need to be cleaned up, hence why the `autoCleanup` option is set to true.

In cases where the replication slots and publications are pre-created, ensure `logrepl.autoCleanup` is set to `false`.